+++ packages/core-js/es/object/has-own.js
@@ -1,0 +1,4 @@
+require('../../modules/es.object.has-own');
+var path = require('../../internals/path');
+
+module.exports = path.Object.hasOwn;