+++ packages/core-js/features/object/index.js
@@ -1,4 +1,5 @@
 var parent = require('../../stable/object');
+// TODO: Remove from `core-js@4`
 require('../../modules/esnext.object.has-own');
 require('../../modules/esnext.object.iterate-entries');
 require('../../modules/esnext.object.iterate-keys');